LayoutTests/fast/scrolling/mac/adjust-scroll-snap-during-gesture-expected.txt


PASS beforeChangeScrollLeft > 0 is true
PASS beforeChangeScrollLeft is scroller.scrollLeft
PASS successfullyParsed is true

TEST COMPLETE
